+++ template/src/navigation/index.tsx
@@ -10,13 +10,17 @@ import { useColorSchemeValue } from '~/hooks/use-color-scheme-value';
 import { useKillswitch } from '~/hooks/use-killswitch';
 import { useSetQueryFocusManager } from '~/hooks/use-set-query-focus-manager';
 import { useSetQueryOnlineManager } from '~/hooks/use-set-query-online-manager';
+import { useInvalidateQueriesOnLanguageChange } from '~/hooks/use-invalidate-queries-on-language-change';
 
 function Navigation() {
   const theme = useColorSchemeValue([DefaultTheme, DarkTheme]);
 
   useKillswitch();
+
   useSetQueryFocusManager();
   useSetQueryOnlineManager();
+
+  useInvalidateQueriesOnLanguageChange();
 
   return (
     <NavigationContainer
